fluid Assert
