+++ python/paddle/jit/api.py
@@ -28,7 +28,11 @@ from paddle.fluid.compiler import (
     CompiledProgram,
     ExecutionStrategy,
 )
-from paddle.fluid.data_feeder import check_type
+from paddle.fluid.data_feeder import (
+    check_variable_and_dtype,
+    check_type,
+)
+from paddle.fluid.layer_helper import LayerHelper
 from paddle.fluid.layers.utils import flatten, pack_sequence_as
 from paddle.fluid.dygraph.base import (
     program_desc_tracing_guard,
@@ -73,6 +77,7 @@ from paddle.fluid.framework import (
 from paddle.fluid.framework import dygraph_only, _non_static_mode
 from paddle.fluid.wrapped_decorator import wrap_decorator
 
+
 __all__ = [
     'TracedLayer',
     'declarative',
@@ -82,6 +87,7 @@ __all__ = [
     'save',
     'load',
     'not_to_static',
+    'Assert',
 ]
 
 
@@ -1834,3 +1840,75 @@ class TracedLayer:
                 params_filename=params_filename,
                 clip_extra=clip_extra,
             )
+
+
+def Assert(cond, data=None, summarize=20, name=None):
+    '''
+    This API creates an op that asserts the given condition is true. If the
+    condition is false, prints the tensors in data. ``summarize`` specifies the
+    number of the elements in the tensors to print.
+
+    Args:
+        cond (Variable): The boolean condition tensor whose numel should be 1.
+        data (list|tuple, optional): list or tuple of tensors to print when
+            condition is not true. If it's ``None``, no tensor will be printed.
+            The default value is ``None``.
+        summarize (int, optional): Number of elements in the tensor to be
+            printed. If its value is -1, then all elements in the tensor will
+            be printed. The default value is 20.
+        name (str, optional): The default value is ``None`` . Normally users
+            don't have to set this parameter. For more information, please
+            refer to :ref:`api_guide_Name` .
+
+    Returns:
+        Operator: the created operation.
+
+    Raises:
+        TypeError: If ``cond`` is not boolean Variable.
+        TypeError: If ``data`` is not a list or tuple or ``None``.
+        TypeError: If ``summarize`` is not int.
+        TypeError: If ``name`` is not a string or ``None`` .
+        framework.core.EnforceNotMet: If the condition is False in running time.
+
+    Examples:
+        .. code-block:: python
+
+            import paddle
+            from paddle.jit.api import Assert
+
+            x = paddle.full(shape=[2, 3], dtype='float32', value=2.0)
+            condition = paddle.max(x) < 1.0 # False
+            Assert(condition, [x], 10, "example_assert_layer")
+
+            exe = paddle.static.Executor()
+            try:
+                exe.run(paddle.static.default_main_program())
+                # Print x and throws paddle.framework.core.EnforceNotMet exception
+                # Example printed message for x:
+                #
+                # Variable: fill_constant_0.tmp_0
+                #   - lod: {}
+                #   - place: CPUPlace()
+                #   - shape: [2, 3]
+                #   - layout: NCHW
+                #   - dtype: float
+                #   - data: [2 2 2 2 2 2]
+            except paddle.framework.core.EnforceNotMet as e:
+                print("Assert Exception Example")
+
+    '''
+    check_variable_and_dtype(cond, "cond", ["bool"], "jit.api.Assert")
+    check_type(data, "data", (list, tuple, type(None)), "jit.api.Assert")
+    check_type(summarize, "summarize", int, "jit.api.Assert")
+    check_type(name, "name", (str, type(None)), "jit.api.Assert")
+
+    layer_name = name if name else ('assert_' + cond.name)
+    helper = LayerHelper(layer_name, **locals())
+
+    op = helper.append_op(
+        type="assert",
+        inputs={"Cond": cond, "Data": [] if data is None else list(data)},
+        attrs={"summarize": summarize},
+    )
+
+    return op
